feat: export the chain config
