+++ src/exports/la-utils/index.ts
@@ -1,6 +1,7 @@
 import { yellowstoneConfig } from "./la-chain/yellowstone/yellowstoneConfig";
 import { contractCall } from "./la-transactions/handlers/contractCall";
 import { sponsoredGasContractCall } from "./la-transactions/handlers/sponsoredGas/sponsoredGasContractCall";
+import { getAlchemyChainConfig } from "./la-helpers/getAlchemyChainConfig";
 import { nativeSend } from "./la-transactions/handlers/nativeSend";
 import { getNonce } from "./la-transactions/primitive/getNonce";
 import { sendTx } from "./la-transactions/primitive/sendTx";
@@ -38,6 +39,7 @@ export const laUtils = {
   },
   helpers: {
     toEthAddress,
+    getAlchemyChainConfig,
   },
 };
