+++ src/templates/tool/package.json
@@ -19,7 +19,7 @@
     "typescript": "^5"
   },
   "dependencies": {
-    "@lit-protocol/vincent-scaffold-sdk": "^1.1.6-alpha.2",
+    "@lit-protocol/vincent-scaffold-sdk": "^1.1.6-alpha.3",
     "@lit-protocol/vincent-tool-sdk": "^1.0.1"
   },
   "exports": {
